code(parser): add support for builtin blocks
These builtin functions are interactive and are things that would
typically require either user interactions or analog components. Right
now, there are four: `toggle`, which creates a single pole double throw
switch, `momentary`, which creates a single pole single throw momentary
switch, `led`, which can be used as an indicator to show ifa wire is
turned on or not, and lastly, `wave`, which while not implemented yet
will generate a square wave of a given frequency / duty cycle.