code: implemnt aworking tflipflop within the go renderer
The state now holds two values: if the flip flop was previously powered (to do rising edge / falling
edge detection) and if the flip flop is on. With this data, the tflipflop code can be runa s many
times in a row as required (ie, if the client renders a frame twice) and not get off. I'm not sure
that the old client limitation could do that.