code: think I finally figured out why the t-flip-flop counters weren't working
The order in which each flip-flop is rendered is important. If an earlier flip flop in the chain is
rendered before a later one in the chain, then the "rising edge" of the later flip flop occurs after
the first flop flop that has been evaluated has already charged it's output wire (in this case, q).

In effect, this means that is n flip flops are all chained together and start off, turning on the
first one would turn on all of them. That's not the desired behavior.

Instead, if we render them in reverse order, we can try to circumvent this. Later flip flops are
rendered first, which means that the rising edges act properly. I don't really like this solution
all that much and think that if could break later if the order of gates change.